Fix (#6)
* refatorando index do TabController

* refatorando o show do TabController

* refatorando index e show de OrderController

* adicionando campos image_url e description no products

* Arrumando o show e index do productsOrdered no OrderController e ProductController

* Completanoo o index e show de TabController, ProductController e OrderController

* packages

* adicionando campos data ao orders

* adicionando mensagens de erro nos controller

* organizando os controllers

* arrumando erro de escrita em orderController

+++ src/controllers/EnvironmentController.js
@@ -2,54 +2,79 @@ const Environment = require('../models/Environment');
 
 class EnvironmentController {
   static async store(req, res) {
-    const { name, is_active } = req.body
-    const environment = await Environment.create({ name, is_active});
+    const { name, is_active } = req.body;
 
-    return res.status(201).json(environment);
+    try {
+      const environment = await Environment.create({ name, is_active });
+
+      return res.status(201).json(environment);
+    } catch (error) {
+      if (error) {
+        return res.status(400).json({ error: 'Try again.' });
+      }
+    }
   }
-  
-  static async index(req, res){
 
-    const environments = await Environment.findAll({include: { association: 'products' }})
+  static async index(req, res) {
+    const environments = await Environment.findAll({
+      include: { association: 'products' },
+    });
 
     return res.status(201).json(environments);
   }
 
-  static async show(req, res){
+  static async show(req, res) {
     const { id } = req.params;
 
     const environment = await Environment.findByPk(id, {
-      include: { association: 'products' }
+      include: { association: 'products' },
     });
+
+    if (environment == null) {
+      return res.status(400).json({ error: 'Environment not found.' });
+    }
 
     return res.json(environment);
   }
 
-  static async update(req, res){
+  static async update(req, res) {
     const { id } = req.params;
     const { body } = req;
 
     await Environment.update(body, {
       where: {
-        id: id
-      }
+        id: id,
+      },
     });
 
-    return res.status(200).json({ message: 'Updated successfully!' });;
+    const environment = await Environment.findByPk(id);
+
+    if (environment == null) {
+      return res.status(400).json({ error: 'Environment not found.' });
+    }
+
+    return res
+      .status(200)
+      .json({ message: 'Updated successfully!', environment });
   }
 
-  static async delete(req, res){
+  static async delete(req, res) {
     const { id } = req.params;
+
+    const environment = await Environment.findByPk(id);
+
+    if (environment == null) {
+      return res.status(400).json({ error: 'Environment not found.' });
+    }
 
     await Environment.destroy({
       where: {
-        id: id
-      }
+        id: id,
+      },
     });
 
-    return res.status(200).json({ message: 'Deleted sucessfully!'});
+    return res.status(200).json({ message: 'Deleted sucessfully!' });
   }
 }
-
 
 module.exports = EnvironmentController;
