+++ src/controllers/ProductController.js
@@ -5,9 +5,15 @@ const Order = require('../models/Order');
 class ProductController {
   static async store(req, res) {
     const { body } = req;
-    const product = await Product.create(body);
+    try {
+      const product = await Product.create(body);
 
-    return res.status(201).json(product);
+      return res.status(201).json(product);
+    } catch (error) {
+      if (error) {
+        return res.status(400).json({ error: 'Try again.' });
+      }
+    }
   }
 
   static async index(req, res) {
@@ -32,21 +38,16 @@ class ProductController {
       ],
     });
 
-    res.status(201).json(products);
+    return res.status(200).json(products);
   }
 
   static async show(req, res) {
     const { id } = req.params;
-
-    const products = await Product.findByPk(id);
-
-    const environment = await Product.findByPk(id, {
+    
+    const product = await Product.findByPk(id, {
       include: [
         {
           association: 'environment',
-          where: {
-            id: products.environment_id,
-          },
         },
         {
           association: 'orders',
@@ -64,7 +65,11 @@ class ProductController {
       attributes: { exclude: ['environment_id'] },
     });
 
-    res.status(201).json(environment);
+    if (product == null) {
+      return res.status(400).json({ error: 'Product not found.' });
+    }
+
+    return res.status(200).json(product);
   }
 
   static async update(req, res) {
@@ -77,19 +82,33 @@ class ProductController {
       },
     });
 
-    return res.status(200).json();
+    const product = await Product.findByPk(id);
+
+    if (product == null) {
+      return res.status(400).json({ error: 'Product not found.' });
+    }
+
+    return res.status(200).json({ message: 'Updated successfully!', product });
   }
 
   static async delete(req, res) {
     const { id } = req.params;
 
+
+    const product = await Product.findByPk(id);
+
+    if (product == null) {
+      return res.status(400).json({ error: 'Product not found.' });
+    }
+    
     await Product.destroy({
       where: {
         id: id,
       },
     });
 
-    return res.status(200).json();
+
+    return res.status(200).json({ message: 'Deleted successfully!' });
   }
 
   static async canSell(id, units) {
